add input forms and components
