+++ src/assets/images/recoverPasswordBg.png
