+++ src/components/shared/inputs/StyledInputLabel.jsx
@@ -1,0 +1,12 @@
+import { InputLabel } from "@mui/material";
+import { styled } from "@mui/material/styles";
+
+const StyledInputLabel = styled(InputLabel)({
+  "&.MuiInputLabel-root": {
+    color: "#2c2e3a",
+    lineHeight: "19px",
+    marginTop: "32px",
+  },
+});
+
+export default StyledInputLabel;
